requirements.txt
grpcio==1.64.0
grpcio-tools==1.64.0
protobuf==5.26.1
pyyaml==6.0.1
SQLAlchemy==2.0.30
pytest==8.2.1
